sysAcad/requirements.txt
asgiref==3.8.1
attrs==23.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-cors-headers==4.3.1
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.0.1
openapi-codec==1.3.2
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zebrafy==1.2.0
